backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
mysqlclient>=2.2.0; sys_platform != 'win32'
pymysql>=1.1.0
cryptography>=42.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
django-filter>=23.5
gTTS>=2.3.0
edge-tts>=7.0.0
deep-translator>=1.11.0
aksharamukha>=2.3
gunicorn>=21.0.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.9
